Implement parseImmigrationCSV() functior
This was quite challenging because the StatsCan CSV was malformed and
some of the columns have trailing numbers and letters that are not
necessary. I used ChatGPT for the regex, as I am not well versed in
regex patterns. I also found a cool way to enable our module to be
directory aware to enable robust file loading.