Add conditional rendering while loading and helper
With the previous implementation chart.js would animate the first 5
items first because the example chart is breifly shown. Now we create
empty items to fill the space while loading so that the animation and
number of values are equivalent! Also added more loading indication and
tested in simulated 3G.